requirements.txt
discord.py
requests
dotenv
python-dotenv
anthropic
mysql-connector-python